fix(specs): make shared MAX_SIGNERS checks type-safe (#176)
## Summary
This PR fixes the shared multisig initialize specs so they keep using
`MAX_SIGNERS` directly while type-checking in both inclusion contexts.

The root issue was that the shared spec is included by both `spl-token`
and `p-token`, but `MAX_SIGNERS` does not have the same type in both
paths. Replacing the old magic numbers with `MAX_SIGNERS` therefore
introduced a type mismatch in the proof/runtime-verification path.

## Testing
- `cargo +nightly-2025-02-16 clippy --manifest-path p-token/Cargo.toml
-Zunstable-options --all-targets --all-features -- --deny=warnings
--deny=clippy::arithmetic_side_effects`
- `cargo +nightly-2025-02-16 check -p spl-token --features
runtime-verification`
- GitHub Actions: `Format & Lint p-token` passed in run `22931335543`

+++ specs/shared/test_process_initialize_multisig2.rs
@@ -38,7 +38,7 @@ fn test_process_initialize_multisig2(
         assert_eq!(result, Err(ProgramError::Custom(0)))
     } else if !((1..=MAX_SIGNERS as usize).contains(&(accounts.len() - 1))) {
         assert_eq!(result, Err(ProgramError::Custom(7)))
-    } else if !(1..=MAX_SIGNERS).contains(&instruction_data[0]) {
+    } else if !((1..=MAX_SIGNERS as usize).contains(&(instruction_data[0] as usize))) {
         assert_eq!(result, Err(ProgramError::Custom(8)))
     } else {
         let multisig_new = get_multisig(&accounts[0]);
